examples/python/requirements.txt
fastapi
searoute
uvicorn
requests
python-dotenv
